+++ src/app/(auth)/forgot-password/page.tsx
@@ -1,72 +1,14 @@
-"use client";
+import ForgetPassword from "@/components/Auth/ForgetPassword";
+import { Suspense } from "react";
 
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import * as z from "zod";
-import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
-import { useForgetPasswordMutation } from "@/hooks/mutations/use-auth-mutations";
-
-const formSchema = z.object({
-  email: z.string().email("Invalid email address"),
-});
-
-export default function ForgotPasswordPage() {
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      email: "",
-    },
-  });
-
-  const { mutate: forgotPassword, isPending} = useForgetPasswordMutation();
-
-  function onSubmit(values: z.infer<typeof formSchema>) {
-    forgotPassword(values);
-  }
-
+const ForgotPasswordPage   = () => {
   return (
-    <div className="flex min-h-screen items-center justify-center">
-      <Card className="w-[400px]">
-        <CardHeader>
-          <CardTitle>Forgot Password</CardTitle>
-          <CardDescription>
-            Enter your email address and we&apos;ll send you a link to reset your password.
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <Form {...form}>
-            <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-              <FormField
-                control={form.control}
-                name="email"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Email</FormLabel>
-                    <FormControl>
-                      <Input placeholder="Enter your email" {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <Button type="submit" className="w-full" disabled={isPending}>
-                {isPending ? "Sending..." : "Send Reset Link"}
-              </Button>
-            </form>
-          </Form>
-        </CardContent>
-      </Card>
-    </div>
+    
+      <Suspense fallback={<div>Loading...</div>}>
+        <ForgetPassword />
+      </Suspense>
+
   );
-} 
+};
+
+export default ForgotPasswordPage;
